Fix: move eventbutton into a seperate component
